move indexer tests to indexer.test.ts
