+++ package.json
@@ -30,7 +30,7 @@
     "build:browser": "node scripts/build-browser.js",
     "test": "vitest run",
     "test:master": "ARK_ENV=master vitest run",
-    "test:unit": "vitest run --exclude test/integration.test.ts",
+    "test:unit": "vitest run --exclude test/integration.test.ts --exclude test/indexer.test.ts",
     "test:setup": "node test/setup.js",
     "test:setup-docker": "node test/setup.js docker",
     "test:build-docker": "docker compose -f docker-compose.yml build --no-cache",
